1-array_iterator.c

+++ function_pointers/1-array_iterator.c
@@ -1,9 +1,14 @@
+#include <stddef.h>
 #include "function_pointers.h"
 /**
+ * array_iterator - array
+ * @array: array
+ * @size: tamanho
+ * @action: accin
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int l;
+	size_t l;
 
 	for (l = 0; l < size; l++)
 	{
